+++ test/tests/test_auxiliary_scripts.py
@@ -40,3 +40,36 @@ def test_rarefaction_curve_0(BTO):
     Odb = pd.read_csv([o for o in outfiles if 'rarefaction_table' in o][0])
     Sdb = pd.read_csv([o for o in outfiles if 'subset_table' in o][0])
     assert set(Odb['samtools_subset'].tolist()) == set(Sdb['samtools_subset'].tolist())
+
+def test_recluster_instrain(BTO):
+    """
+    Just run the program on two settings to make sure the -a arguments works
+    """
+    # Run program
+    outfile = BTO.test_dir + 'test.tsv'
+    sl = test_utils.get_aux_script_loc('recluster_instrain_compare')
+    db_loc = os.path.join(BTO.RC_Loc, 'output/RC_test_genomeWide_compare.tsv')
+
+    cmd = f"{sl} --version"
+    print(cmd)
+    call(cmd, shell=True)
+
+    cmd = f"{sl} -d {db_loc} -o {outfile} -a 99"
+    print(cmd)
+    call(cmd, shell=True)
+
+    # Load output
+    outfiles = glob.glob(outfile + '*')
+    assert len(outfiles) == 1
+    Odb = pd.read_csv(outfiles[0], sep='\t')
+    assert len(Odb['cluster'].unique()) == 2
+
+    cmd = f"{sl} -d {db_loc} -o {outfile} -a 99.995"
+    print(cmd)
+    call(cmd, shell=True)
+
+    # Load output
+    outfiles = glob.glob(outfile + '*')
+    assert len(outfiles) == 1
+    Odb = pd.read_csv(outfiles[0], sep='\t')
+    assert len(Odb['cluster'].unique()) == 3